+++ scripts/gas-refund-program/staking/2.0/AbstractStateTracker.ts
